+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added "deleteKey" and "deleteAllKeys" methods to the URL Manager
 
 ## [0.0.5] - 2026-02-24
 - Removed console.log from updateKey method of url manager